+++ sql-plugin/src/main/spark321db/scala/org/apache/spark/rapids/shims/GpuShuffleExchangeExec.scala
@@ -31,5 +31,5 @@ case class GpuShuffleExchangeExec(
     child: SparkPlan,
     shuffleOrigin: ShuffleOrigin)(
     cpuOutputPartitioning: Partitioning)
-  extends GpuShuffleExchangeExecBase(gpuOutputPartitioning,
+  extends GpuDatabricksShuffleExchangeExecBase(gpuOutputPartitioning,
     child, shuffleOrigin)(cpuOutputPartitioning)
